requirements.txt
Flask==2.3.2
azure-data-tables
azure-servicebus